requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
gspread
oauth2client
pandas
scikit-learn
numpy